+++ StreamChatSwiftUITests/Tests/ChatChannel/MessageListViewAvatars_Tests.swift
@@ -86,6 +86,7 @@ class MessageListViewAvatars_Tests: StreamChatTestCase {
             currentDateString: nil,
             listId: "listId",
             isMessageThread: false,
+            shouldShowTypingIndicator: false,
             onMessageAppear: { _ in },
             onScrollToBottom: {},
             onLongPress: { _ in }