+++ CHANGELOG.md
@@ -3,6 +3,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [4.21.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.21.0)
+_September 02, 2022_
+
 ### ✅ Added
 - Configuring avatars visibility in groups
 - Method to swap the `MessageRepliesView`